requirements.txt
discord.py==2.0.0
requests==2.32.2
python-dotenv==0.19.2
